[730] removed useless cast in Benchmark class.
Eclipse warnings really seem to have an actual purpose besides painting your code yellow, in the end.

+++ java/src/moonaframework/util/Benchmark.java
@@ -16,7 +16,7 @@ public final class Benchmark {
 	}
 	public static void showTime(Snippet s) throws NullArgumentException {
 		long time = Benchmark.time(s);
-		System.out.println(time + " ns | " + ((double) time) * 10e-10 + " s");
+		System.out.println(time + " ns | " + (time) * 10e-10 + " s");
 	}
 	
 	public static double stress(int cycles, Snippet s) throws IllegalArgumentException, NullArgumentException {
